Esok present

+++ src/main/webapp/donationList.jsp
@@ -13,23 +13,26 @@
 </head>
 <body>
 <sql:setDataSource var="ic" driver="org.postgresql.Driver"
-				   url="jdbc:postgresql://ec2-3-228-235-79.compute-1.amazonaws.com/ddrev47ip327l0"
-				   user = "hlcietwdsgkwyq"
-				   password="f6078446e3932c85a4d99b3753e1b04295a6add4a27ee4fdc3649c1efb1a04f1"/>
+                   url="jdbc:postgresql://ec2-3-228-235-79.compute-1.amazonaws.com/ddrev47ip327l0"
+                   user = "hlcietwdsgkwyq"
+                   password="f6078446e3932c85a4d99b3753e1b04295a6add4a27ee4fdc3649c1efb1a04f1"/>
 <sql:query dataSource="${ic}" var="oc">
-	SELECT row_number() over () "rank",donationID,donationName,donationDescription,donationOpenDate, donationCloseDate from donation where donationId>0
+	SELECT donationID, donationName, donationDescription, TO_CHAR(donationOpenDate, 'dd/MM/YYYY')donationOpenDate, TO_CHAR(donationCloseDate, 'dd/MM/YYYY') donationCloseDate FROM donation where donationID>0
 </sql:query>
-<nav>
-	<ul>
-		<div class = "logo">
-			<img src = "Image/Logo.png" alt = "logo">
-			<span class="add">Masjid Sultan Azlan Shah, Jalan Raja Musa Mahadi, Institut Tadbiran Islam Perak, 31400 Ipoh, Perak, Malaysia 05-5456779</span>
-			<li><a class="navi" href="#" onclick="document.location='LoginAccount.html'">LOG KELUAR</a>
-			<li><a class="navi" href="#" onclick="document.location='ViewAccount.html' ">AKAUN</a>
-			<li><a class="navi" href="#" onclick="document.location='buttonApplicationCommittee.html'">PERMOHONAN</a>
-			<li><a class="navi" href="#" onclick="document.location='donationList.jsp' ">BANTUAN</a>
-		</div>
-	</ul>
+   <nav>
+        <div class = "logo">
+            <img src = "Image/Logo.png" alt = "logo">
+            <span class="add">Masjid Sultan Azlan Shah, Jalan Raja Musa Mahadi, Institut Tadbiran Islam Perak, 31400 Ipoh, Perak, Malaysia 05-5456779</span>
+            
+            <form class="bar" method="get" action="CommitteeHandler">
+		        <input type="hidden" name="action" value="logoutCommittee">
+		       	<button class="navi" href="#" onclick="return confirm('Anda pasti untuk log keluar?');">LOG KELUAR</button>
+		    </form>
+            
+            <button class="navi" href="#" onclick="document.location='committeeAccount.jsp'">AKAUN</button>
+            <button class="navi" href="#" onclick="document.location='applicationType.jsp'">PERMOHONAN</button>
+            <button class="navi" href="#"><i>BANTUAN</i></button>
+        </div>
 </nav>
 <a class="tambah" href=donationForm.jsp>TAMBAH</a>
 <br>
@@ -39,7 +42,7 @@
 	<div style="overflow-x:auto;">
 		<table style="text-align: center;" id="donationList">
 			<tr>
-				<th onclick="sortTable(0)" style="width: 50px; height: 50px;">NO. </th>
+				<th onclick="sortTable(0)" style="width: 50px; height: 50px;">DONATION ID</th>
 				<th onclick="sortTable(1)" style="width: 250px;">NAMA BANTUAN</th>
 				<th onclick="sortTable(2)" style="width: 500px;">MAKLUMAT BANTUAN</th>
 				<th onclick="sortTable(3)" style="width: 150px;">TARIKH BUKA</th>
@@ -49,7 +52,7 @@
 			<c:forEach var="result" items="${oc.rows}">
 				<tr>
 					<td class="donationID">
-						<c:out value="${result.rank}"/>
+						<c:out value="${result.donationID}"/>
 					</td>
 					<td class="donationName">
 						<c:out value="${result.donationName}"/>
@@ -71,7 +74,7 @@
 						<form method="post">
 							<input type="hidden" name="donationID" value="${result.donationID}">
 							<input type="hidden" name="action" value="deleteDonation">
-							<button  class="action"  type="delete" formaction="DonationHandler">PADAM</button>
+							<button  class="action"  type="delete" formaction="DonationHandler" onclick="return confirm('Anda pasti untuk membuang bantuan ini?');">PADAM</button>
 						</form>
 					</td>
 				</tr>
@@ -96,49 +99,60 @@
 	nav a.navi:hover{
 		margin-top: 10px;
 		color: #BBE9FE;}
-	.add {
-		position: absolute;
-		left: 90px;
-		width: 270px;
-		border: none;
-		padding: 20px 16px;;
-		font-size: 14px;
-		color: #455a64;
-		font-family: "Bahnschrift", "Courier New", monospace;}
-	.logo{
-		flex: 1;}
-	.logo a{
-		display: flex;
-		align-items: center;
-		font-size: 20px;
-	}
-	.logo img{
-		width: 100px;
-		height: 100px;
-		margin-right: 10px;
-	}
-	ul {
-		list-style-type: none;
-		overflow: hidden;
-		background-color: #BBE9FE;
-		height: 100px;
-		width:100%;}
-	li {
-		float: right;}
-	.navi {
-		display: block;
-		color: #455a64;
-		font-weight: bold;
-		text-align: center;
-		padding: 30px 16px;
-		text-decoration: none;
-		margin-top: 10px;
-		margin-right: 10px;
-	}
-	li a.navi:hover {
-		background-color: #617C8D;
-		height: 5px;
-	}
+	       /*css for navigation bar*/
+     .add {
+     	position: absolute;
+	    left: 90px;
+	    width: 270px;
+	    border: none;
+	    padding: 20px 16px;;
+	    font-size: 14px;
+	    color: #455a64;
+	    font-family: "Bahnschrift", "Courier New", monospace;}
+	    
+	    nav {
+	    	background-color: #b3e6ff;
+	    }
+	    
+	    nav a.navi:hover{
+	        margin-top: 10px;
+	        color: #BBE9FE;}
+	    .logo{
+	        flex: 1;}
+	    .logo button {
+	        display: flex;
+	        align-items: center;
+	        font-size: 20px;
+	    }
+	    .logo img{
+	        width: 100px;
+	        height: 100px;
+	        margin-right: 10px;
+	    }
+	    button {
+	        float: right;}
+	    .navi {
+	        display: block;
+	        color: #455a64;
+	        font-weight: bold;
+	        text-align: center;
+	        padding: 30px 16px;
+	        text-decoration: none;
+	        margin-top: 10px;
+	        margin-right: 10px;
+	        border: none;
+	        background-color: #b3e6ff;
+	    }
+	    
+	    .bar {
+	    	display: inline;
+	    }
+	    
+	     button.navi:hover {
+	        background-color: #617C8D;
+	        height: 5px;
+	    }
+		/*css for navigation bar*/
 	div.a {
 		position: relative;
 		display: block;
